fix(math): pipeline blocking due to calling no-op rewards/adv
Currently in pipeline mode, reward and advantages are computed by the
rollout.

However, to align with collocated mode, the actor's
compute_rewards and compute_advantages_and_returns are still invoked.

This does not introduce computation overhead as all computations are
skipped if rewards and advantages already exist in the rollout results.
So the actor simply acquires a result and puts it into the output
channel.

However, since both the compute_rewards and run_training functions
belong to the actor group, and we do not allow concurrent execution of
different functions of the same group, this effectively blocks training
until all rollout results (from inference/rollout) have been get by
compute_rewards and put into channel.

This hotfix only temporarily resolves the issue by avoiding any get-put
in compute_rewards and compute_advantages_and_returns if in pipeline
mode.
An ideal fix should make reward an independent entity (worker group) and
remove reward computation from rollout.
This should be done in the next step when we align SGLangWorker and
AsyncSGLangWorker.

Signed-off-by: Hao Lin <linhaomails@gmail.com>